corrigindo bugs e controle de acesso

+++ listarCategorias/listar.php
@@ -2,10 +2,23 @@
 require_once "../config.php";
 require_once "../head.php";
 
+session_start();
+
+$login = $_SESSION['login'];
+
+if (!isset($login)) {
+    header('Location: index.php');
+}
+
+$sql = "select nivel from usuarios where email = '$login' and status = 'Ativo'";
+$buscar = mysqli_query($connection, $sql);
+$array = mysqli_fetch_array($buscar);
+$nivel = $array['nivel'];
 ?>
+
 <body>
 
-    <a class="btn btn-primary" style="margin: 20px 0 0 140px;" href="../landing/landing.php" role="button"><i class="fas fa-undo"></i>&nbsp;Voltar</a>
+    <a class="btn btn-primary" style="margin: 20px 0 0 140px;" href="../menu.php" role="button"><i class="fas fa-undo"></i>&nbsp;Voltar</a>
     <div class="container" style="margin-top: 20px; width: 500px;">
         <h3 style="padding: 20px; text-align: center;">Categorias</h3>
 
@@ -18,7 +31,7 @@ require_once "../head.php";
             </thead>
 
             <?php
-            
+
             $sql = "SELECT * FROM `categorias`;";
             $query = mysqli_query($connection, $sql);
 
@@ -32,13 +45,16 @@ require_once "../head.php";
                     <td>
                         <a class="btn btn-warning btn-sm" href="editar.php?id=<?php echo $id; ?>" role="button"><i class="far fa-edit"></i>&nbsp;Editar</a>
 
-                        <a class="btn btn-danger btn-sm" href="delete.php?id=<?php echo $id; ?>" role="button"><i class="far fa-trash-alt"></i>&nbsp;Exluir</a>
+                        <?php if ($nivel == 1) { ?>
+                            <a class="btn btn-danger btn-sm" href="delete.php?id=<?php echo $id; ?>" role="button"><i class="far fa-trash-alt"></i>&nbsp;Exluir</a>
+                        <?php } ?>
+
                     </td>
                 <?php } ?>
                 </tr>
         </table>
         <div style="text-align: right; margin-bottom: 30px;">
-            
+
             <a class="btn btn-success right" href="cadastrar.php" role="button"><i class="fas fa-plus-circle"></i>&nbsp;Adicionar Novo</a>
         </div>
     </div>
